pyproject.toml
[project]
name = "danger-rose"
version = "0.1.0"
description = ""
authors = [
    {name = "Samuel Vange",email = "7166607+svange@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
augint-github = "^1.3.1"
pytest = "^8.4.1"
pytest-html = "^4.1.1"
pytest-cov = "^6.2.1"
black = "^25.1.0"
ruff = "^0.12.3"
pre-commit = "^4.2.0"


[tool.poetry.dependencies]
pygame-ce = "^2.5.5"
